requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Financial Data & APIs
alpha-vantage>=2.3.1
yfinance>=0.2.18
pandas-datareader>=0.10.0
requests>=2.31.0

# NLP & Sentiment Analysis
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.12.0
nltk>=3.8.1
spacy>=3.6.0
textblob>=0.17.1
vaderSentiment>=3.3.2

# Reinforcement Learning
gymnasium>=0.28.0
stable-baselines3>=2.0.0
tensorboard>=2.13.0

# Data Processing & Analysis
scipy>=1.10.0
statsmodels>=0.14.0
arch>=6.2.0
ta>=0.10.2
empyrical>=0.5.5

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Research & Paper Generation
jinja2>=3.1.0
markdown>=3.4.0
beautifulsoup4>=4.12.0

# Performance & Optimization
numba>=0.57.0
cython>=3.0.0

# Database & Storage
sqlalchemy>=2.0.0
h5py>=3.9.0

# Jupyter & Interactive Development
jupyter>=1.0.0
ipykernel>=6.24.0
ipywidgets>=8.0.0
notebook>=7.0.0
